Add files via upload
First off I'm not sure why I did not commit these individually as I was going. Things have been crazy and I must have forgot. Second, most of it was readings and so there is not a ton of deliverables. However the 25 hour PyTorch tutorial was completed and that is presented in workbook form. Due to lack of time at end of the semester I did no compelte every single exercise and extra curriculum for each of the topics however I did look them over.